Correctness issue solved + some barrier removal
The correctness issue was caused by the reduction on `rho`
rho += r[j]*r[j]. The solution includes removing the clause
and using a `#pragma omp atomic` for mutual exclusion.

TODO: Remove argo::barriers in conj_grad, which have to do
with the data structure updates protected by the Argo locks.
Use selective coherence to reduce the overhead of synchronization.

NOTE: Transitioning from the reduction to the atomic in the
relevant code block, a decrease of about 400 Mop/s was observed.