Pad ArgoDSM global structures and eliminate write misses
This commit pads by `pad` the chunks to be worked on by the nodes.
The default size of `pad` is 512 (1 page since we work on doubles).

The way the nodes touch memory (regular accesses) looks like this:
| node0's chunk | pad | node1's chunk | pad | node2's chunk | ...

We also pad `gnorms`, which is used to accumulate the scalars:
| node0's scalars  | node1's scalars  | node2's scalars | ...
<------ pad ------><------ pad ------><------ pad ------>

! This commit is not thoroughly performance tested.
Home checkout commit: 970850d4b11a6ca2bf33852282464ca827197105