Fix corner case in CG
Fix corner case when calculating `chunks_home` for the `z` and `p`
arrays (irregular accesses). The bug appeared on Karolina running
on 32 nodes for CLASS=C.
Issue: `chunks_home` was getting a value `> numtasks-1` after some
values in the last nodes' chunk.

Performance (referring to the previous commit):
The performance of the previous commit was evaluated for up to 16
nodes on the Karolina cluster.
1 : 305.30sec
2 : 172.70sec
4 : 109.72sec
8 :  75.45sec
16:  69.03sec

! Important:
- The performance evaluation corresponding to the above measureme-
nts was done running 1 process and 1 thread per node as there is
an issue of multiple write misses appearing when running multiple
threads with OpenMP (ArgoDSM related issue).
- The memory allocation policy used for the performance evaluation
was first-touch. It is very important to use a policy for placing
the pages exactly where you want them, and as such avoid write mi-
sses. First-touch saves a lot of effort in achieving that. However,
execution seems to hang for 32 nodes (ArgoDSM related issue).