Add distribution and distribution type option in NAS CG
The distribution option `DISTRIBUTION_TYPE` is added.

DISTRIBUTION_TYPE = `0` (default)
- This is the simplest distribution that eliminates all Argo
write misses. In this distribution we simply work with a pa-
ge-aligned chunk in the `distribute` function.

DISTRIBUTION_TYPE = '1'
- In this distribution we pad each nodes' chunk to be page
aligned. With this distribution we preserve better load ba-
lance between the nodes by adding a bit of memory footprint.

Both distributions remain to be tested!